+++ database/migrations/2021_04_12_104406_add_price_smallest.php
@@ -14,7 +14,7 @@ class AddPriceSmallest extends Migration
     public function up()
     {
         Schema::table('buddhists', function (Blueprint $table) {
-            $table->integer('priceSmallest');
+            $table->float('priceSmallest');
         });
     }
